add tests for creating label filter
